Simplify schema builder
