+++ examples/getstarted/components/test-a-zdazd/test-edit-3.json
@@ -1,0 +1,14 @@
+{
+  "connection": "default",
+  "collectionName": "components_test_a_zdazd_test_edit_3_s",
+  "info": {
+    "name": "Test edit 3",
+    "icon": "test",
+    "description": "IONADZI\nazdinazodin"
+  },
+  "attributes": {
+    "key": {
+      "type": "string"
+    }
+  }
+}
