+++ packages/strapi-plugin-content-type-builder/services/schema-builder/content-type-builder.js
