+++ packages/strapi-plugin-content-type-builder/services/schema-builder/index.js
@@ -1,0 +1,120 @@
+'use strict';
+
+const path = require('path');
+
+const createSchemaHandler = require('./schema-handler');
+const createComponentBuilder = require('./component-builder');
+const createContentTypeBuilder = require('./content-type-builder');
+
+module.exports = function createBuilder() {
+  const components = Object.keys(strapi.components).map(key => {
+    const compo = strapi.components[key];
+
+    return {
+      modelName: compo.modelName,
+      plugin: compo.modelName,
+      uid: compo.uid,
+      filename: compo.__filename__,
+      dir: path.join(strapi.dir, 'components', compo.category),
+      schema: compo.__schema__,
+    };
+  });
+
+  const contentTypes = Object.keys(strapi.contentTypes).map(key => {
+    const contentType = strapi.contentTypes[key];
+
+    let dir;
+    if (contentType.plugin) {
+      dir = `./extensions/${contentType.plugin}/models`;
+    } else {
+      dir = `./api/${contentType.apiName}/models`;
+    }
+
+    return {
+      modelName: contentType.modelName,
+      plugin: contentType.plugin,
+      uid: contentType.uid,
+      filename: contentType.__filename__,
+      dir: path.join(strapi.dir, dir),
+      schema: contentType.__schema__,
+    };
+  });
+
+  return createSchemaBuilder({
+    components,
+    contentTypes,
+  });
+};
+
+/**
+ * Schema builder
+ */
+function createSchemaBuilder({ components, contentTypes }) {
+  const tmpComponents = new Map();
+  const tmpContentTypes = new Map();
+
+  // init temporary ContentTypes
+  Object.keys(contentTypes).forEach(key => {
+    tmpContentTypes.set(
+      contentTypes[key].uid,
+      createSchemaHandler(contentTypes[key])
+    );
+  });
+
+  // init temporary components
+  Object.keys(components).forEach(key => {
+    tmpComponents.set(
+      components[key].uid,
+      createSchemaHandler(components[key])
+    );
+  });
+
+  return {
+    get components() {
+      return tmpComponents;
+    },
+    get contentTypes() {
+      return tmpContentTypes;
+    },
+
+    ...createComponentBuilder({ tmpComponents, tmpContentTypes }),
+    ...createContentTypeBuilder({ tmpComponents, tmpContentTypes }),
+
+    /**
+     * Write all type to files
+     */
+    writeFiles() {
+      return Promise.all(
+        [
+          ...Array.from(tmpComponents.values()),
+          ...Array.from(tmpContentTypes.values()),
+        ].map(schema => schema.flush())
+      )
+        .catch(error => {
+          strapi.log.error('Error writing schema files');
+          strapi.log.error(error);
+          return this.rollback();
+        })
+        .catch(error => {
+          strapi.log.error(
+            'Error rolling back schema files. You might need to fix your files manually'
+          );
+          strapi.log.error(error);
+
+          throw new Error('Invalid schema edition');
+        });
+    },
+
+    /**
+     * rollback all files
+     */
+    rollback() {
+      return Promise.all(
+        [
+          ...Array.from(tmpComponents.values()),
+          ...Array.from(tmpContentTypes.values()),
+        ].map(schema => schema.rollback())
+      );
+    },
+  };
+}
