+++ packages/strapi-plugin-content-type-builder/services/schema-builder/schema-handler.js
